+++ log/error.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
-import { getLogger } from "./_get_logger.ts";
+import { getLogger } from "./get_logger.ts";
 import { type GenericFunction } from "./logger.ts";
 
 /** Log with error level, using default logger. */
